bump Node runtime version from 12 to 16
Upgrade the required Node runtime version from the current 12 to the GitHub-recommended 16. This removes the deprecation warnings that started to appear when using this action on GitHub (`Node.js 12 actions are deprecated. For more information see: https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/. Please update the following actions to use Node.js 16: falti/dotenv-action`)